move dynamic_bitset code out of variant.hpp and other small changes
