+++ libraries/libfc/test/CMakeLists.txt
@@ -14,6 +14,7 @@ add_executable( test_fc
         scoped_exit/test_scoped_exit.cpp
         static_variant/test_static_variant.cpp
         variant/test_variant.cpp
+        variant/test_variant_dynamic_bitset.cpp
         variant_estimated_size/test_variant_estimated_size.cpp
         test_base64.cpp
         test_escape_str.cpp
